Introduce InitConfigAction
Action places the default configuration file (packaged with the app) in the user's
home directory. In the future, the app will prefer configuration at this location
over the packaged default.

This change also introduces an injectable logger, which allows for better testing
of the log messages. Testing that output is getting logged is important since
it is the main vehicle of communcation to the user here.