Restructure Actions as Subcommands
These commands are distinct. The command line option seemed to make sense
initially, but implementing the interface was awkward because it always required
passing 5 arguments that I didn't care about at all. Now that each subcommand
sets the "cmd" variable in the namespace, it's easy to resolve a runnable action
to be executed.