Only Record New Builds
Unfortunately Ektorp uses Jackson 1.9 while everything else uses Jackson 2.3.
Ideally Ektorp would be migrated to the latest supported version so we could
avoid duplicated dependencies.

DateTimes are now serialized in the human-readable ISO format. Duration has
been changed to durationInMillis and is stored as a long, not a Duration object.